+++ src/SwagMigrationMagento.php
@@ -25,11 +25,12 @@ class SwagMigrationMagento extends Plugin
 
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/DependencyInjection/'));
         $loader->load('magento.xml');
-        $loader->load('magento19.xml');
+        // $loader->load('magento19.xml');
         $loader->load('magento20.xml');
         $loader->load('magento21.xml');
         $loader->load('magento22.xml');
         $loader->load('magento23.xml');
+        $loader->load('magento24.xml');
     }
 
     public function rebuildContainer(): bool